+++ testproject/settings.py
@@ -31,7 +31,7 @@ INSTALLED_APPS = [
     'django.contrib.sites',
     'django.contrib.staticfiles',
     'testapp',
-    'bootstrap4',
+    'django_bootstrap5',
 ]
 
 MIDDLEWARE = [